refactor(project): small improvements and refactor in Serial executor
