+++ project/.gitignore
@@ -1,10 +1,12 @@
-### SOFTWARE FACTORY ###
-# TODO: should be removed from gitignore when start to edit them
+### Unibo Software Factory ###
 sites/
-bin/
+# src/**/*.java
+# src-gen/
+# srcMore/
+# build_*.gradle
 
-# Created by https://www.gitignore.io/api/java,xtext,gradle,eclipse,intellij
-# Edit at https://www.gitignore.io/?templates=java,xtext,gradle,eclipse,intellij
+# Created by https://www.gitignore.io/api/java,xtext,gradle,eclipse,intellij+all
+# Edit at https://www.gitignore.io/?templates=java,xtext,gradle,eclipse,intellij+all
 
 ### Eclipse ###
 .metadata
@@ -75,7 +77,7 @@ local.properties
 
 .sts4-cache/
 
-### Intellij ###
+### Intellij+all ###
 # Covers JetBrains IDEs: IntelliJ, RubyMine, PhpStorm, AppCode, PyCharm, CLion, Android Studio and WebStorm
 # Reference: https://intellij-support.jetbrains.com/hc/en-us/articles/206544839
 
@@ -85,10 +87,6 @@ local.properties
 .idea/**/usage.statistics.xml
 .idea/**/dictionaries
 .idea/**/shelf
-
-# Markdown Navigator
-.idea/markdown-navigator.xml
-.idea/markdown-navigator
 
 # Generated files
 .idea/**/contentModel.xml
@@ -149,13 +147,18 @@ fabric.properties
 # Android studio 3.1+ serialized cache file
 .idea/caches/build_file_checksums.ser
 
-### Intellij Patch ###
-# Comment Reason: https://github.com/joeblau/gitignore.io/issues/186#issuecomment-215987721
+### Intellij+all Patch ###
+# Ignores the whole .idea folder and all .iml files
+# See https://github.com/joeblau/gitignore.io/issues/186 and https://github.com/joeblau/gitignore.io/issues/360
 
-# *.iml
-# modules.xml
-# .idea/misc.xml
-# *.ipr
+.idea/
+
+# Reason: https://github.com/joeblau/gitignore.io/issues/186#issuecomment-249601023
+
+*.iml
+modules.xml
+.idea/misc.xml
+*.ipr
 
 # Sonarlint plugin
 .idea/sonarlint
@@ -213,4 +216,4 @@ gradle-app.setting
 ### Gradle Patch ###
 **/build/
 
-# End of https://www.gitignore.io/api/java,xtext,gradle,eclipse,intellij
+# End of https://www.gitignore.io/api/java,xtext,gradle,eclipse,intellij+all
