--- project/.idea/codeStyles/Project.xml
@@ -1,86 +1,0 @@
-<component name="ProjectCodeStyleConfiguration">
-  <code_scheme name="Project" version="173">
-    <JavaCodeStyleSettings>
-      <option name="GENERATE_FINAL_LOCALS" value="true" />
-      <option name="CLASS_COUNT_TO_USE_IMPORT_ON_DEMAND" value="99" />
-      <option name="NAMES_COUNT_TO_USE_IMPORT_ON_DEMAND" value="99" />
-      <option name="PACKAGES_TO_USE_IMPORT_ON_DEMAND">
-        <value />
-      </option>
-      <option name="IMPORT_LAYOUT_TABLE">
-        <value>
-          <package name="" withSubpackages="true" static="true" />
-          <emptyLine />
-          <package name="" withSubpackages="true" static="false" />
-        </value>
-      </option>
-      <option name="JD_ADD_BLANK_AFTER_PARM_COMMENTS" value="true" />
-      <option name="JD_ADD_BLANK_AFTER_RETURN" value="true" />
-      <option name="JD_DO_NOT_WRAP_ONE_LINE_COMMENTS" value="true" />
-    </JavaCodeStyleSettings>
-    <JetCodeStyleSettings>
-      <option name="PACKAGES_TO_USE_STAR_IMPORTS">
-        <value>
-          <package name="tornadofx" withSubpackages="false" static="false" />
-        </value>
-      </option>
-      <option name="NAME_COUNT_TO_USE_STAR_IMPORT" value="2147483647" />
-      <option name="NAME_COUNT_TO_USE_STAR_IMPORT_FOR_MEMBERS" value="2147483647" />
-      <option name="CONTINUATION_INDENT_IN_PARAMETER_LISTS" value="true" />
-      <option name="CONTINUATION_INDENT_IN_ARGUMENT_LISTS" value="true" />
-      <option name="CONTINUATION_INDENT_FOR_EXPRESSION_BODIES" value="true" />
-      <option name="CONTINUATION_INDENT_FOR_CHAINED_CALLS" value="true" />
-      <option name="CONTINUATION_INDENT_IN_SUPERTYPE_LISTS" value="true" />
-      <option name="CONTINUATION_INDENT_IN_IF_CONDITIONS" value="true" />
-      <option name="CONTINUATION_INDENT_IN_ELVIS" value="true" />
-      <option name="WRAP_EXPRESSION_BODY_FUNCTIONS" value="0" />
-      <option name="IF_RPAREN_ON_NEW_LINE" value="false" />
-      <option name="CODE_STYLE_DEFAULTS" value="KOTLIN_OFFICIAL" />
-    </JetCodeStyleSettings>
-    <MarkdownNavigatorCodeStyleSettings>
-      <option name="WRAP_ON_TYPING" value="0" />
-      <option name="RIGHT_MARGIN" value="72" />
-    </MarkdownNavigatorCodeStyleSettings>
-    <ScalaCodeStyleSettings>
-      <option name="classCountToUseImportOnDemand" value="99" />
-      <option name="NOT_CONTINUATION_INDENT_FOR_PARAMS" value="true" />
-      <option name="USE_ALTERNATE_CONTINUATION_INDENT_FOR_PARAMS" value="true" />
-      <option name="SD_BLANK_LINE_AFTER_PARAMETERS_COMMENTS" value="true" />
-      <option name="SD_BLANK_LINE_AFTER_RETURN_COMMENTS" value="true" />
-      <option name="SD_BLANK_LINE_BEFORE_PARAMETERS" value="true" />
-    </ScalaCodeStyleSettings>
-    <codeStyleSettings language="JAVA">
-      <option name="KEEP_SIMPLE_BLOCKS_IN_ONE_LINE" value="true" />
-      <option name="KEEP_SIMPLE_METHODS_IN_ONE_LINE" value="true" />
-      <option name="KEEP_SIMPLE_LAMBDAS_IN_ONE_LINE" value="true" />
-      <option name="KEEP_SIMPLE_CLASSES_IN_ONE_LINE" value="true" />
-      <arrangement>
-        <groups>
-          <group>
-            <type>GETTERS_AND_SETTERS</type>
-            <order>KEEP</order>
-          </group>
-          <group>
-            <type>OVERRIDDEN_METHODS</type>
-            <order>KEEP</order>
-          </group>
-        </groups>
-      </arrangement>
-    </codeStyleSettings>
-    <codeStyleSettings language="Scala">
-      <option name="ALIGN_MULTILINE_PARAMETERS" value="false" />
-    </codeStyleSettings>
-    <codeStyleSettings language="kotlin">
-      <option name="CODE_STYLE_DEFAULTS" value="KOTLIN_OFFICIAL" />
-      <option name="CALL_PARAMETERS_WRAP" value="0" />
-      <option name="CALL_PARAMETERS_LPAREN_ON_NEXT_LINE" value="false" />
-      <option name="CALL_PARAMETERS_RPAREN_ON_NEXT_LINE" value="false" />
-      <option name="METHOD_PARAMETERS_WRAP" value="0" />
-      <option name="METHOD_PARAMETERS_LPAREN_ON_NEXT_LINE" value="false" />
-      <option name="METHOD_PARAMETERS_RPAREN_ON_NEXT_LINE" value="false" />
-      <option name="EXTENDS_LIST_WRAP" value="0" />
-      <option name="METHOD_CALL_CHAIN_WRAP" value="0" />
-      <option name="ASSIGNMENT_WRAP" value="0" />
-    </codeStyleSettings>
-  </code_scheme>
-</component>
